fix(monolith): move savings cache TTL test off cross-loop asyncio.run()
CI caught a real bug: bazel_core_test.py's cached_bazel_query_savings
tests called asyncio.run() twice per test, each spinning up its own
event loop. The module-level asyncio.Lock() binds to whichever loop
first acquires it, so the second asyncio.run() call in
test_savings_endpoint_refetches_after_ttl_expires hit a cross-loop
lock error and the read never happened (calls["n"] stayed 0).

Move that coverage to bazel_router_test.py, which already drives the
same cache through one persistent event loop via TestClient (the
pattern the postgres demo's savings tests already use for exactly
this reason). Drop the duplicate direct-call cache tests from
bazel_core_test.py in favor of the router-level ones.
